Make field $args more specific & flexible
Dear @aristath,
I am currently working on a Module for Kirki that requires setting `'input_attrs'` specific for a control. Currently when settings it e.g. for a multicolor field the same value is copied to each color control. That makes the following scenario impossible.

````
'choices'   => [
	'menu_item'   => esc_attr__( 'Menu Item Color', 'sage' ),
	'menu_item_hover' => esc_attr__( 'Menu Item Hover & Active', 'sage' ),
],
'input_attrs' => [
	'menu_item_hover' => [
		'data-my-test' => 'The data-my-test value',
	],
],
```

This PR adds a function that allows defining `'input_attrs'` for all controls at once (as it is currently the case) but also allows to set it specific for a single control as in my example above. I wrote it to be usable also with other fields and changed it in the PR also for the `'default'` `$arg`. This allows (vice versa) to define the default for all control at once or for each separately.

Looking forward to read your feedback,

Philipp