Adds solution to define the "alpha" option for single controls
The functionality of the PR can be used to allow setting the "alpha" setting on a per control basis, too. Currently I did not find a way to allow setting the alpha option for controls within Multicolor Field as it requires setting a `choices` argument to `alpha` what simply add another color control with the title "alpha". With this additional PR the following  is possible:

`````
'choices'   => [
	'menu_item'   => esc_attr__( 'Menu Item Color', 'domain' ),
	'menu_item_hover' => esc_attr__( 'Menu Item Hover & Active', 'domain' ),
],
'alpha' => [
	'menu_item' => true,
],
````
what only add the alpha selection to the control `menu_item`.

OR

`````
'alpha' => true,
`````
what adds the alpha selection to all controls of the multicolor field.